requirements.txt
pytest~=7.1.2
selenium~=4.2.0
PyHamcrest~=2.0.3
